divide coponenet style
